add untested view to create a bulk purchase invitation
git-svn-id: https://repos.nextthought.com/svn/nti-svn/nti.app.store/trunk@56090 3dfb626b-079f-4ab1-a9f6-4654e5377508

+++ src/nti/app/store/utils.py
@@ -49,7 +49,7 @@ def is_valid_amount(amount):
 
 def is_valid_pve_int(value):
 	try:
-		value = float(value)
+		value = int(value)
 		return value > 0
 	except (TypeError, ValueError):
 		return False
